searchbot/requirements.txt
httpx>=0.27
beautifulsoup4>=4.12
openai>=1.30
rich>=13.7
python-dotenv>=1.0
prompt_toolkit>=3.0
selenium>=4.20
